see 11/15/17 update notes in ReadMe.md
**11/15/17 Update**
- Renamed StateTester class to StateDemo to more accurately indicate that this is a demonstration of how the provided State pattern can be implemented.
- Changed ButtonContext.buttonState visibility to protected and removed redundant class variable ButtonAContext.buttonState.
- Provided default functionality in ButtonContext methods and removed abstract designation.
- Updated javadoc and readme to reflect these changes.
