+++ Documentation/javadoc/allclasses-frame.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:11 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:36 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>All Classes (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>All Classes (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
 <script type="text/javascript" src="script.js"></script>
 </head>
@@ -19,7 +19,7 @@
 <li><a href="cs505/group1/state/StateA.html" title="class in cs505.group1.state" target="classFrame">StateA</a></li>
 <li><a href="cs505/group1/state/StateB.html" title="class in cs505.group1.state" target="classFrame">StateB</a></li>
 <li><a href="cs505/group1/state/StateC.html" title="class in cs505.group1.state" target="classFrame">StateC</a></li>
-<li><a href="cs505/group1/state/StateTester.html" title="class in cs505.group1.state" target="classFrame">StateTester</a></li>
+<li><a href="cs505/group1/state/StateDemo.html" title="class in cs505.group1.state" target="classFrame">StateDemo</a></li>
 </ul>
 </div>
 </body>
