+++ Documentation/javadoc/cs505/group1/state/ButtonState.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:10 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>ButtonState (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>ButtonState (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="ButtonState (ButtonState Javadoc)";
+            parent.document.title="ButtonState (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
@@ -106,25 +106,19 @@ var activeTableTab = "activeTableTab";
 <hr>
 <br>
 <pre>public interface <span class="typeNameLabel">ButtonState</span></pre>
-<div class="block">A set of methods for GrovePi button interactions as a function of machine state. 
+<div class="block">A set of methods for GrovePi button interactions which depend on machine state. 
     
  Button actions include singlePress, doublePress, and longPress.  Additional actions
- can be defined for specific requirements.  <br>
+ can be defined for specific requirements.<br>
  <br>
- The button methods set the state within the context.  
- This is accomplished by returning an object of the desired concrete state 
- which has been typecast to the more generic ButtonState.
- (ie. <i> return (ButtonState) new StateA(); </i> in the examples)</div>
+ Button press methods can return a different state by setting the returned object
+ to any implemented concrete state cast to ButtonState.</div>
 <dl>
 <dt><span class="simpleTagLabel">Version:</span></dt>
-<dd>2017.10.29</dd>
+<dd>2017.11.15</dd>
 <dt><span class="simpleTagLabel">Author:</span></dt>
 <dd>CS505 Fall17 Group1: <br>
- Emily Park<br>
- Jeffrey Blankenship  <br>
- Cecelia Oluwadoyinsola<br>
- James Luczynski<br>
- Melissa Mulcahy <br></dd>
+ Emily Park, Jeffrey Blankenship, Cecelia Oluwadoyinsola, James Luczynski, Melissa Mulcahy <br></dd>
 </dl>
 </li>
 </ul>
