+++ Documentation/javadoc/cs505/group1/state/StateB.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:10 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>StateB (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>StateB (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="StateB (ButtonState Javadoc)";
+            parent.document.title="StateB (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
@@ -118,15 +118,8 @@ extends java.lang.Object
 implements <a href="../../../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a></pre>
 <div class="block">Example of a concrete subclass of ButtonState</div>
 <dl>
-<dt><span class="simpleTagLabel">Version:</span></dt>
-<dd>2017.10.29</dd>
 <dt><span class="simpleTagLabel">Author:</span></dt>
-<dd>CS505 Fall17 Group1: <br>
- Emily Park<br>
- Jeffrey Blankenship  <br>
- Cecelia Oluwadoyinsola<br>
- James Luczynski<br>
- Melissa Mulcahy <br></dd>
+<dd>CS505 Fall17 Group1</dd>
 </dl>
 </li>
 </ul>
@@ -168,19 +161,19 @@ implements <a href="../../../cs505/group1/state/ButtonState.html" title="interfa
 <tr id="i0" class="altColor">
 <td class="colFirst"><code><a href="../../../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a></code></td>
 <td class="colLast"><code><span class="memberNameLink"><a href="../../../cs505/group1/state/StateB.html#doublePress--">doublePress</a></span>()</code>
-<div class="block">Processes a double button press, with capability to change state.</div>
+<div class="block">Calls doublePress method of the concrete subclass.</div>
 </td>
 </tr>
 <tr id="i1" class="rowColor">
 <td class="colFirst"><code><a href="../../../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a></code></td>
 <td class="colLast"><code><span class="memberNameLink"><a href="../../../cs505/group1/state/StateB.html#longPress--">longPress</a></span>()</code>
-<div class="block">Processes a long button press, with capability to change state.</div>
+<div class="block">Calls longPress method of the concrete subclass.</div>
 </td>
 </tr>
 <tr id="i2" class="altColor">
 <td class="colFirst"><code><a href="../../../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a></code></td>
 <td class="colLast"><code><span class="memberNameLink"><a href="../../../cs505/group1/state/StateB.html#singlePress--">singlePress</a></span>()</code>
-<div class="block">Processes a single button press, with capability to change state.</div>
+<div class="block">Calls singlePress method of the concrete subclass.</div>
 </td>
 </tr>
 <tr id="i3" class="rowColor">
@@ -236,12 +229,13 @@ implements <a href="../../../cs505/group1/state/ButtonState.html" title="interfa
 <li class="blockList">
 <h4>singlePress</h4>
 <pre>public&nbsp;<a href="../../../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a>&nbsp;singlePress()</pre>
-<div class="block">Processes a single button press, with capability to change state.</div>
+<div class="block"><span class="descfrmTypeLabel">Description copied from interface:&nbsp;<code><a href="../../../cs505/group1/state/ButtonState.html#singlePress--">ButtonState</a></code></span></div>
+<div class="block">Calls singlePress method of the concrete subclass.</div>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
 <dd><code><a href="../../../cs505/group1/state/ButtonState.html#singlePress--">singlePress</a></code>&nbsp;in interface&nbsp;<code><a href="../../../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a></code></dd>
 <dt><span class="returnLabel">Returns:</span></dt>
-<dd>ButtonState object</dd>
+<dd>a ButtonState object set by the concrete subclass.</dd>
 </dl>
 </li>
 </ul>
@@ -252,12 +246,13 @@ implements <a href="../../../cs505/group1/state/ButtonState.html" title="interfa
 <li class="blockList">
 <h4>doublePress</h4>
 <pre>public&nbsp;<a href="../../../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a>&nbsp;doublePress()</pre>
-<div class="block">Processes a double button press, with capability to change state.</div>
+<div class="block"><span class="descfrmTypeLabel">Description copied from interface:&nbsp;<code><a href="../../../cs505/group1/state/ButtonState.html#doublePress--">ButtonState</a></code></span></div>
+<div class="block">Calls doublePress method of the concrete subclass.</div>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
 <dd><code><a href="../../../cs505/group1/state/ButtonState.html#doublePress--">doublePress</a></code>&nbsp;in interface&nbsp;<code><a href="../../../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a></code></dd>
 <dt><span class="returnLabel">Returns:</span></dt>
-<dd>ButtonState object</dd>
+<dd>a ButtonState object set by the concrete subclass.</dd>
 </dl>
 </li>
 </ul>
@@ -268,12 +263,13 @@ implements <a href="../../../cs505/group1/state/ButtonState.html" title="interfa
 <li class="blockList">
 <h4>longPress</h4>
 <pre>public&nbsp;<a href="../../../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a>&nbsp;longPress()</pre>
-<div class="block">Processes a long button press, with capability to change state.</div>
+<div class="block"><span class="descfrmTypeLabel">Description copied from interface:&nbsp;<code><a href="../../../cs505/group1/state/ButtonState.html#longPress--">ButtonState</a></code></span></div>
+<div class="block">Calls longPress method of the concrete subclass.</div>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
 <dd><code><a href="../../../cs505/group1/state/ButtonState.html#longPress--">longPress</a></code>&nbsp;in interface&nbsp;<code><a href="../../../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a></code></dd>
 <dt><span class="returnLabel">Returns:</span></dt>
-<dd>ButtonState object</dd>
+<dd>a ButtonState object set by the concrete subclass.</dd>
 </dl>
 </li>
 </ul>
