+++ Documentation/javadoc/cs505/group1/state/class-use/StateB.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:10 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Uses of Class cs505.group1.state.StateB (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>Uses of Class cs505.group1.state.StateB (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Uses of Class cs505.group1.state.StateB (ButtonState Javadoc)";
+            parent.document.title="Uses of Class cs505.group1.state.StateB (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
