+++ Documentation/javadoc/cs505/group1/state/class-use/StateDemo.html
@@ -4,7 +4,7 @@
 <head>
 <!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Uses of Class cs505.group1.state.ButtonContext (ButtonState (CS505 Group1))</title>
+<title>Uses of Class cs505.group1.state.StateDemo (ButtonState (CS505 Group1))</title>
 <meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../script.js"></script>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Uses of Class cs505.group1.state.ButtonContext (ButtonState (CS505 Group1))";
+            parent.document.title="Uses of Class cs505.group1.state.StateDemo (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
@@ -33,7 +33,7 @@
 </a>
 <ul class="navList" title="Navigation">
 <li><a href="../../../../cs505/group1/state/package-summary.html">Package</a></li>
-<li><a href="../../../../cs505/group1/state/ButtonContext.html" title="class in cs505.group1.state">Class</a></li>
+<li><a href="../../../../cs505/group1/state/StateDemo.html" title="class in cs505.group1.state">Class</a></li>
 <li class="navBarCell1Rev">Use</li>
 <li><a href="../../../../overview-tree.html">Tree</a></li>
 <li><a href="../../../../deprecated-list.html">Deprecated</a></li>
@@ -47,8 +47,8 @@
 <li>Next</li>
 </ul>
 <ul class="navList">
-<li><a href="../../../../index.html?cs505/group1/state/class-use/ButtonContext.html" target="_top">Frames</a></li>
-<li><a href="ButtonContext.html" target="_top">No&nbsp;Frames</a></li>
+<li><a href="../../../../index.html?cs505/group1/state/class-use/StateDemo.html" target="_top">Frames</a></li>
+<li><a href="StateDemo.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
 <li><a href="../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
@@ -70,36 +70,9 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h2 title="Uses of Class cs505.group1.state.ButtonContext" class="title">Uses of Class<br>cs505.group1.state.ButtonContext</h2>
+<h2 title="Uses of Class cs505.group1.state.StateDemo" class="title">Uses of Class<br>cs505.group1.state.StateDemo</h2>
 </div>
-<div class="classUseContainer">
-<ul class="blockList">
-<li class="blockList">
-<ul class="blockList">
-<li class="blockList"><a name="cs505.group1.state">
-<!--   -->
-</a>
-<h3>Uses of <a href="../../../../cs505/group1/state/ButtonContext.html" title="class in cs505.group1.state">ButtonContext</a> in <a href="../../../../cs505/group1/state/package-summary.html">cs505.group1.state</a></h3>
-<table class="useSummary" border="0" cellpadding="3" cellspacing="0" summary="Use table, listing subclasses, and an explanation">
-<caption><span>Subclasses of <a href="../../../../cs505/group1/state/ButtonContext.html" title="class in cs505.group1.state">ButtonContext</a> in <a href="../../../../cs505/group1/state/package-summary.html">cs505.group1.state</a></span><span class="tabEnd">&nbsp;</span></caption>
-<tr>
-<th class="colFirst" scope="col">Modifier and Type</th>
-<th class="colLast" scope="col">Class and Description</th>
-</tr>
-<tbody>
-<tr class="altColor">
-<td class="colFirst"><code>class&nbsp;</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../../cs505/group1/state/ButtonAContext.html" title="class in cs505.group1.state">ButtonAContext</a></span></code>
-<div class="block">An example extension of ButtonContext.</div>
-</td>
-</tr>
-</tbody>
-</table>
-</li>
-</ul>
-</li>
-</ul>
-</div>
+<div class="classUseContainer">No usage of cs505.group1.state.StateDemo</div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
 <div class="bottomNav"><a name="navbar.bottom">
 <!--   -->
@@ -110,7 +83,7 @@
 </a>
 <ul class="navList" title="Navigation">
 <li><a href="../../../../cs505/group1/state/package-summary.html">Package</a></li>
-<li><a href="../../../../cs505/group1/state/ButtonContext.html" title="class in cs505.group1.state">Class</a></li>
+<li><a href="../../../../cs505/group1/state/StateDemo.html" title="class in cs505.group1.state">Class</a></li>
 <li class="navBarCell1Rev">Use</li>
 <li><a href="../../../../overview-tree.html">Tree</a></li>
 <li><a href="../../../../deprecated-list.html">Deprecated</a></li>
@@ -124,8 +97,8 @@
 <li>Next</li>
 </ul>
 <ul class="navList">
-<li><a href="../../../../index.html?cs505/group1/state/class-use/ButtonContext.html" target="_top">Frames</a></li>
-<li><a href="ButtonContext.html" target="_top">No&nbsp;Frames</a></li>
+<li><a href="../../../../index.html?cs505/group1/state/class-use/StateDemo.html" target="_top">Frames</a></li>
+<li><a href="StateDemo.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
 <li><a href="../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
