+++ Documentation/javadoc/cs505/group1/state/package-summary.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:10 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>cs505.group1.state (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>cs505.group1.state (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="cs505.group1.state (ButtonState Javadoc)";
+            parent.document.title="cs505.group1.state (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
@@ -71,11 +71,6 @@
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
 <h1 title="Package" class="title">Package&nbsp;cs505.group1.state</h1>
-<div class="docSummary">
-<div class="block">This package provides developer tools to code a Finite State Machine using a
- a GrovePi button sensor.</div>
-</div>
-<p>See:&nbsp;<a href="#package.description">Description</a></p>
 </div>
 <div class="contentContainer">
 <ul class="blockList">
@@ -90,7 +85,7 @@
 <tr class="altColor">
 <td class="colFirst"><a href="../../../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a></td>
 <td class="colLast">
-<div class="block">A set of methods for GrovePi button interactions as a function of machine state.</div>
+<div class="block">A set of methods for GrovePi button interactions which depend on machine state.</div>
 </td>
 </tr>
 </tbody>
@@ -107,13 +102,13 @@
 <tr class="altColor">
 <td class="colFirst"><a href="../../../cs505/group1/state/ButtonAContext.html" title="class in cs505.group1.state">ButtonAContext</a></td>
 <td class="colLast">
-<div class="block">An example concrete subclass of ButtonContext.</div>
+<div class="block">An example extension of ButtonContext.</div>
 </td>
 </tr>
 <tr class="rowColor">
 <td class="colFirst"><a href="../../../cs505/group1/state/ButtonContext.html" title="class in cs505.group1.state">ButtonContext</a></td>
 <td class="colLast">
-<div class="block">An abstract framework for developing and using contexts in a finite state machine
+<div class="block">A framework for developing and using contexts in a finite state machine
  to set actions generated by a GrovePi button sensor.</div>
 </td>
 </tr>
@@ -136,34 +131,15 @@
 </td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a href="../../../cs505/group1/state/StateTester.html" title="class in cs505.group1.state">StateTester</a></td>
+<td class="colFirst"><a href="../../../cs505/group1/state/StateDemo.html" title="class in cs505.group1.state">StateDemo</a></td>
 <td class="colLast">
-<div class="block">Example of a Finite State Machine using the provided concrete subclass examples.</div>
+<div class="block">Demonstration of a Finite State Machine using the provided concrete subclass examples.</div>
 </td>
 </tr>
 </tbody>
 </table>
 </li>
 </ul>
-<a name="package.description">
-<!--   -->
-</a>
-<h2 title="Package cs505.group1.state Description">Package cs505.group1.state Description</h2>
-<div class="block">This package provides developer tools to code a Finite State Machine using a
- a GrovePi button sensor. The button sensor can be used to change the state of the machine
- and/or to execute functions which change with the current state of the machine.  Functionality
- is provided to set the state of the machine by method call as well.</div>
-<dl>
-<dt><span class="simpleTagLabel">Version:</span></dt>
-<dd>2017.10.29</dd>
-<dt><span class="simpleTagLabel">Author:</span></dt>
-<dd>CS505 Fall17 Group1: <br>
- Emily Park<br>
- Jeffrey Blankenship  <br>
- Cecelia Oluwadoyinsola<br>
- James Luczynski<br>
- Melissa Mulcahy <br></dd>
-</dl>
 </div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
 <div class="bottomNav"><a name="navbar.bottom">
