+++ Documentation/javadoc/cs505/group1/state/package-use.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:11 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Uses of Package cs505.group1.state (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>Uses of Package cs505.group1.state (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Uses of Package cs505.group1.state (ButtonState Javadoc)";
+            parent.document.title="Uses of Package cs505.group1.state (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
@@ -85,13 +85,13 @@
 <tbody>
 <tr class="altColor">
 <td class="colOne"><a href="../../../cs505/group1/state/class-use/ButtonContext.html#cs505.group1.state">ButtonContext</a>
-<div class="block">An abstract framework for developing and using contexts in a finite state machine
+<div class="block">A framework for developing and using contexts in a finite state machine
  to set actions generated by a GrovePi button sensor.</div>
 </td>
 </tr>
 <tr class="rowColor">
 <td class="colOne"><a href="../../../cs505/group1/state/class-use/ButtonState.html#cs505.group1.state">ButtonState</a>
-<div class="block">A set of methods for GrovePi button interactions as a function of machine state.</div>
+<div class="block">A set of methods for GrovePi button interactions which depend on machine state.</div>
 </td>
 </tr>
 </tbody>
