+++ Documentation/javadoc/deprecated-list.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:11 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:36 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Deprecated List (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>Deprecated List (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
 <script type="text/javascript" src="script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Deprecated List (ButtonState Javadoc)";
+            parent.document.title="Deprecated List (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
