+++ Documentation/javadoc/index-files/index-1.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:11 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>B-Index (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>B-Index (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <script type="text/javascript" src="../script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="B-Index (ButtonState Javadoc)";
+            parent.document.title="B-Index (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
@@ -76,22 +76,28 @@
 <dl>
 <dt><a href="../cs505/group1/state/ButtonAContext.html" title="class in cs505.group1.state"><span class="typeNameLink">ButtonAContext</span></a> - Class in <a href="../cs505/group1/state/package-summary.html">cs505.group1.state</a></dt>
 <dd>
-<div class="block">An example concrete subclass of ButtonContext.</div>
+<div class="block">An example extension of ButtonContext.</div>
 </dd>
 <dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonAContext.html#ButtonAContext-cs505.group1.state.ButtonState-">ButtonAContext(ButtonState)</a></span> - Constructor for class cs505.group1.state.<a href="../cs505/group1/state/ButtonAContext.html" title="class in cs505.group1.state">ButtonAContext</a></dt>
-<dd>&nbsp;</dd>
+<dd>
+<div class="block">Constructs a new ButtonAContext object and initializes it with the specified ButtonState</div>
+</dd>
 <dt><a href="../cs505/group1/state/ButtonContext.html" title="class in cs505.group1.state"><span class="typeNameLink">ButtonContext</span></a> - Class in <a href="../cs505/group1/state/package-summary.html">cs505.group1.state</a></dt>
 <dd>
-<div class="block">An abstract framework for developing and using contexts in a finite state machine
+<div class="block">A framework for developing and using contexts in a finite state machine
  to set actions generated by a GrovePi button sensor.</div>
 </dd>
 <dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonContext.html#ButtonContext-cs505.group1.state.ButtonState-">ButtonContext(ButtonState)</a></span> - Constructor for class cs505.group1.state.<a href="../cs505/group1/state/ButtonContext.html" title="class in cs505.group1.state">ButtonContext</a></dt>
 <dd>
 <div class="block">Constructs a newly allocated ButtonContext object that represents the specified buttonState.</div>
 </dd>
+<dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonContext.html#buttonState">buttonState</a></span> - Variable in class cs505.group1.state.<a href="../cs505/group1/state/ButtonContext.html" title="class in cs505.group1.state">ButtonContext</a></dt>
+<dd>
+<div class="block">This instance variable stores the current buttonState.</div>
+</dd>
 <dt><a href="../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state"><span class="typeNameLink">ButtonState</span></a> - Interface in <a href="../cs505/group1/state/package-summary.html">cs505.group1.state</a></dt>
 <dd>
-<div class="block">A set of methods for GrovePi button interactions as a function of machine state.</div>
+<div class="block">A set of methods for GrovePi button interactions which depend on machine state.</div>
 </dd>
 </dl>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">L</a>&nbsp;<a href="index-6.html">M</a>&nbsp;<a href="index-7.html">S</a>&nbsp;<a href="index-8.html">T</a>&nbsp;</div>
