+++ Documentation/javadoc/index-files/index-2.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:11 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>C-Index (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>C-Index (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <script type="text/javascript" src="../script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="C-Index (ButtonState Javadoc)";
+            parent.document.title="C-Index (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
@@ -75,10 +75,7 @@
 <h2 class="title">C</h2>
 <dl>
 <dt><a href="../cs505/group1/state/package-summary.html">cs505.group1.state</a> - package cs505.group1.state</dt>
-<dd>
-<div class="block">This package provides developer tools to code a Finite State Machine using a
- a GrovePi button sensor.</div>
-</dd>
+<dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">L</a>&nbsp;<a href="index-6.html">M</a>&nbsp;<a href="index-7.html">S</a>&nbsp;<a href="index-8.html">T</a>&nbsp;</div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
