+++ Documentation/javadoc/index-files/index-4.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:11 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>G-Index (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>G-Index (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <script type="text/javascript" src="../script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="G-Index (ButtonState Javadoc)";
+            parent.document.title="G-Index (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
@@ -74,8 +74,6 @@
 </a>
 <h2 class="title">G</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonAContext.html#getState--">getState()</a></span> - Method in class cs505.group1.state.<a href="../cs505/group1/state/ButtonAContext.html" title="class in cs505.group1.state">ButtonAContext</a></dt>
-<dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonContext.html#getState--">getState()</a></span> - Method in class cs505.group1.state.<a href="../cs505/group1/state/ButtonContext.html" title="class in cs505.group1.state">ButtonContext</a></dt>
 <dd>
 <div class="block">Gets the current state.</div>
