+++ Documentation/javadoc/index-files/index-5.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:11 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>L-Index (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>L-Index (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <script type="text/javascript" src="../script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="L-Index (ButtonState Javadoc)";
+            parent.document.title="L-Index (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
@@ -74,11 +74,10 @@
 </a>
 <h2 class="title">L</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonAContext.html#longPress--">longPress()</a></span> - Method in class cs505.group1.state.<a href="../cs505/group1/state/ButtonAContext.html" title="class in cs505.group1.state">ButtonAContext</a></dt>
-<dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonContext.html#longPress--">longPress()</a></span> - Method in class cs505.group1.state.<a href="../cs505/group1/state/ButtonContext.html" title="class in cs505.group1.state">ButtonContext</a></dt>
 <dd>
-<div class="block">Implements code for a long button press.</div>
+<div class="block">Invokes the longPress method of the concrete class for the
+ current state.</div>
 </dd>
 <dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonState.html#longPress--">longPress()</a></span> - Method in interface cs505.group1.state.<a href="../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a></dt>
 <dd>
@@ -89,13 +88,9 @@
 <div class="block">Processes a long button press, with capability to change state.</div>
 </dd>
 <dt><span class="memberNameLink"><a href="../cs505/group1/state/StateB.html#longPress--">longPress()</a></span> - Method in class cs505.group1.state.<a href="../cs505/group1/state/StateB.html" title="class in cs505.group1.state">StateB</a></dt>
-<dd>
-<div class="block">Processes a long button press, with capability to change state.</div>
-</dd>
+<dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../cs505/group1/state/StateC.html#longPress--">longPress()</a></span> - Method in class cs505.group1.state.<a href="../cs505/group1/state/StateC.html" title="class in cs505.group1.state">StateC</a></dt>
-<dd>
-<div class="block">Processes a long button press, with capability to change state.</div>
-</dd>
+<dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">L</a>&nbsp;<a href="index-6.html">M</a>&nbsp;<a href="index-7.html">S</a>&nbsp;<a href="index-8.html">T</a>&nbsp;</div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
