+++ Documentation/javadoc/index-files/index-6.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:11 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>M-Index (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>M-Index (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <script type="text/javascript" src="../script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="M-Index (ButtonState Javadoc)";
+            parent.document.title="M-Index (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
@@ -74,7 +74,7 @@
 </a>
 <h2 class="title">M</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../cs505/group1/state/StateTester.html#main-java.lang.String:A-">main(String[])</a></span> - Static method in class cs505.group1.state.<a href="../cs505/group1/state/StateTester.html" title="class in cs505.group1.state">StateTester</a></dt>
+<dt><span class="memberNameLink"><a href="../cs505/group1/state/StateDemo.html#main-java.lang.String:A-">main(String[])</a></span> - Static method in class cs505.group1.state.<a href="../cs505/group1/state/StateDemo.html" title="class in cs505.group1.state">StateDemo</a></dt>
 <dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">L</a>&nbsp;<a href="index-6.html">M</a>&nbsp;<a href="index-7.html">S</a>&nbsp;<a href="index-8.html">T</a>&nbsp;</div>
