+++ Documentation/javadoc/index-files/index-7.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:11 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>S-Index (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>S-Index (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
 <script type="text/javascript" src="../script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="S-Index (ButtonState Javadoc)";
+            parent.document.title="S-Index (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
@@ -74,17 +74,14 @@
 </a>
 <h2 class="title">S</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonAContext.html#setState-cs505.group1.state.ButtonState-">setState(ButtonState)</a></span> - Method in class cs505.group1.state.<a href="../cs505/group1/state/ButtonAContext.html" title="class in cs505.group1.state">ButtonAContext</a></dt>
-<dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonContext.html#setState-cs505.group1.state.ButtonState-">setState(ButtonState)</a></span> - Method in class cs505.group1.state.<a href="../cs505/group1/state/ButtonContext.html" title="class in cs505.group1.state">ButtonContext</a></dt>
 <dd>
 <div class="block">Sets the state to be equal to the passed state.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonAContext.html#singlePress--">singlePress()</a></span> - Method in class cs505.group1.state.<a href="../cs505/group1/state/ButtonAContext.html" title="class in cs505.group1.state">ButtonAContext</a></dt>
-<dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonContext.html#singlePress--">singlePress()</a></span> - Method in class cs505.group1.state.<a href="../cs505/group1/state/ButtonContext.html" title="class in cs505.group1.state">ButtonContext</a></dt>
 <dd>
-<div class="block">Implements code for a single button press.</div>
+<div class="block">Invokes the singlePress method of the concrete class for the
+ current state.</div>
 </dd>
 <dt><span class="memberNameLink"><a href="../cs505/group1/state/ButtonState.html#singlePress--">singlePress()</a></span> - Method in interface cs505.group1.state.<a href="../cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a></dt>
 <dd>
@@ -95,13 +92,9 @@
 <div class="block">Processes a single button press, with capability to change state.</div>
 </dd>
 <dt><span class="memberNameLink"><a href="../cs505/group1/state/StateB.html#singlePress--">singlePress()</a></span> - Method in class cs505.group1.state.<a href="../cs505/group1/state/StateB.html" title="class in cs505.group1.state">StateB</a></dt>
-<dd>
-<div class="block">Processes a single button press, with capability to change state.</div>
-</dd>
+<dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../cs505/group1/state/StateC.html#singlePress--">singlePress()</a></span> - Method in class cs505.group1.state.<a href="../cs505/group1/state/StateC.html" title="class in cs505.group1.state">StateC</a></dt>
-<dd>
-<div class="block">Processes a single button press, with capability to change state.</div>
-</dd>
+<dd>&nbsp;</dd>
 <dt><a href="../cs505/group1/state/StateA.html" title="class in cs505.group1.state"><span class="typeNameLink">StateA</span></a> - Class in <a href="../cs505/group1/state/package-summary.html">cs505.group1.state</a></dt>
 <dd>
 <div class="block">Example of a concrete subclass of ButtonState</div>
@@ -126,11 +119,11 @@
 <dd>
 <div class="block">Constructs a newly allocated StateC object.</div>
 </dd>
-<dt><a href="../cs505/group1/state/StateTester.html" title="class in cs505.group1.state"><span class="typeNameLink">StateTester</span></a> - Class in <a href="../cs505/group1/state/package-summary.html">cs505.group1.state</a></dt>
+<dt><a href="../cs505/group1/state/StateDemo.html" title="class in cs505.group1.state"><span class="typeNameLink">StateDemo</span></a> - Class in <a href="../cs505/group1/state/package-summary.html">cs505.group1.state</a></dt>
 <dd>
-<div class="block">Example of a Finite State Machine using the provided concrete subclass examples.</div>
+<div class="block">Demonstration of a Finite State Machine using the provided concrete subclass examples.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../cs505/group1/state/StateTester.html#StateTester--">StateTester()</a></span> - Constructor for class cs505.group1.state.<a href="../cs505/group1/state/StateTester.html" title="class in cs505.group1.state">StateTester</a></dt>
+<dt><span class="memberNameLink"><a href="../cs505/group1/state/StateDemo.html#StateDemo--">StateDemo()</a></span> - Constructor for class cs505.group1.state.<a href="../cs505/group1/state/StateDemo.html" title="class in cs505.group1.state">StateDemo</a></dt>
 <dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">L</a>&nbsp;<a href="index-6.html">M</a>&nbsp;<a href="index-7.html">S</a>&nbsp;<a href="index-8.html">T</a>&nbsp;</div>
