+++ Documentation/javadoc/overview-tree.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_144) on Wed Nov 01 02:24:11 EDT 2017 -->
+<!-- Generated by javadoc (1.8.0_144) on Thu Nov 16 00:09:35 EST 2017 -->
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Class Hierarchy (ButtonState Javadoc)</title>
-<meta name="date" content="2017-11-01">
+<title>Class Hierarchy (ButtonState (CS505 Group1))</title>
+<meta name="date" content="2017-11-16">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
 <script type="text/javascript" src="script.js"></script>
 </head>
@@ -13,7 +13,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Class Hierarchy (ButtonState Javadoc)";
+            parent.document.title="Class Hierarchy (ButtonState (CS505 Group1))";
         }
     }
     catch(err) {
@@ -89,7 +89,7 @@
 <li type="circle">cs505.group1.state.<a href="cs505/group1/state/StateA.html" title="class in cs505.group1.state"><span class="typeNameLink">StateA</span></a> (implements cs505.group1.state.<a href="cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a>)</li>
 <li type="circle">cs505.group1.state.<a href="cs505/group1/state/StateB.html" title="class in cs505.group1.state"><span class="typeNameLink">StateB</span></a> (implements cs505.group1.state.<a href="cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a>)</li>
 <li type="circle">cs505.group1.state.<a href="cs505/group1/state/StateC.html" title="class in cs505.group1.state"><span class="typeNameLink">StateC</span></a> (implements cs505.group1.state.<a href="cs505/group1/state/ButtonState.html" title="interface in cs505.group1.state">ButtonState</a>)</li>
-<li type="circle">cs505.group1.state.<a href="cs505/group1/state/StateTester.html" title="class in cs505.group1.state"><span class="typeNameLink">StateTester</span></a></li>
+<li type="circle">cs505.group1.state.<a href="cs505/group1/state/StateDemo.html" title="class in cs505.group1.state"><span class="typeNameLink">StateDemo</span></a></li>
 </ul>
 </li>
 </ul>
