+++ Source/cs505/group1/state/ButtonContext.java
@@ -1,21 +1,22 @@
 package cs505.group1.state;
 
 /**
- * An abstract framework for developing and using contexts in a finite state machine
+ * A framework for developing and using contexts in a finite state machine
  * to set actions generated by a GrovePi button sensor.
+ * This class can be used as provided for a single context, or extended into 
+ * multiple contexts for additional flexibility. 
+ * 
+ * 
  * @author CS505 Fall17 Group1: <br>
- * Emily Park<br>
- * Jeffrey Blankenship  <br>
- * Cecelia Oluwadoyinsola<br>
- * James Luczynski<br>
- * Melissa Mulcahy <br>
- * @version 2017.10.29
+ * Emily Park, Jeffrey Blankenship, Cecelia Oluwadoyinsola, James Luczynski, Melissa Mulcahy <br>
+ * @version 2017.11.15
  */
-public abstract class ButtonContext {
+public class ButtonContext {
  
-    //No javadoc provided: private visibility.
-    //This instance variable stores the current buttonState.
-    private ButtonState buttonState;
+    /**
+     * This instance variable stores the current buttonState.
+     */
+    protected ButtonState buttonState;
     
     /**
      * Constructs a newly allocated ButtonContext object that represents the specified buttonState.
@@ -27,36 +28,52 @@ public abstract class ButtonContext {
     }
     
     /**
-     * Implements code for a single button press.
+     * Invokes the singlePress method of the concrete class for the
+     * current state.  
      */
-    public abstract void singlePress();
+    public void singlePress(){
+      this.buttonState.singlePress();
+    };
     
     /**
-     * Implements code for a double button press.
+     * Invokes the doublePress method of the concrete class for the
+     * current state.  
      */
-    public abstract void doublePress();
+    public void doublePress(){
+      buttonState = buttonState.doublePress();
+    };
     
     /**
-     * Implements code for a long button press.
+     * Invokes the longPress method of the concrete class for the
+     * current state.  
      */
-    public abstract void longPress();
+    public void longPress(){
+      buttonState = buttonState.longPress();
+    };
     
     /**
      * Sets the state to be equal to the passed state.
      * @param buttonState The state to be set.
      */
-    public abstract void setState(ButtonState  buttonState);
+    public void setState(ButtonState  buttonState){
+      this.buttonState = buttonState;
+    };
     
     /**
      * Gets the current state.
      * @return ButtonState
      */
-    public abstract ButtonState getState();
+    public ButtonState getState(){
+      return this.buttonState;
+    };
     
     /**
      * Invokes the toString method of the concrete class for the
      * current state.  
+     * @return String
      */
-    public abstract String toString();
-    
+    @Override
+    public String toString(){
+        return this.buttonState.toString();
+    };
 }
