--- Source/cs505/group1/state/README.md
@@ -1,11 +1,0 @@
-# Button Interface
-
------
-Usage
------
-
-  This is a java interface for usage with the button device that connects to
-the Raspberry Pi.  It follows a state design pattern and is customizable for most
-desired uses.
-
-NOTE: StateA.java, StateB.java, StateC.java, and ButtonAContext.java are samples on how to utilize the abstract class and interface provided. StateTester.java demonstrates a way to implement this.
